+++ cwd.txt
@@ -1,0 +1,63 @@
+st_Dashboard/
+    .env
+    .gitattributes
+    .gitignore
+    convert_env_to_toml.py
+    LICENSE
+    prompt.txt
+    quick_scripts.ipynb
+    README.md
+    requirements.txt
+    streamlit_app.py
+    .streamlit/
+        secrets.toml
+    custom_moduls/
+        Connection_handling.py
+        calender_widget/
+            helper_functions.py
+            name_template.html
+            title_template.html
+            __init__.py
+            __pycache__/
+                helper_functions.cpython-311.pyc
+                __init__.cpython-311.pyc
+        Prompter/
+            __init__.py
+            __pycache__/
+                __init__.cpython-311.pyc
+        streamlit_timeline/
+            __init__.py
+            frontend/
+                build/
+                    asset-manifest.json
+                    bootstrap.min.css
+                    index.html
+                    index_bk.html
+                    precache-manifest.a632b80b25b19a605a90dd282a0397aa.js
+                    service-worker.js
+                    static/
+                        css/
+                            2.ab85016c.chunk.css
+                            2.ab85016c.chunk.css.map
+                            styles.css
+                        js/
+                            2.fc3cfd4b.chunk.js
+                            2.fc3cfd4b.chunk.js.LICENSE.txt
+                            2.fc3cfd4b.chunk.js.map
+                            main.46d91172.chunk.js
+                            main.46d91172.chunk.js.map
+                            runtime-main.44d30fc2.js
+                            runtime-main.44d30fc2.js.map
+            styles/
+                background.css
+            __pycache__/
+                __init__.cpython-311.pyc
+        __pycache__/
+            Connection_handling.cpython-311.pyc
+    static/
+        styles.css
+        assets/
+            Logo.png
+    views/
+        about_us.py
+        tamam_calender.py
