+++ streamlit_app.py
@@ -1,5 +1,5 @@
 import streamlit as st
-from custom_moduls.Connection_handling import Connection_Handler
+from modules.Connection_handling import Connection_Handler
 
 st.set_page_config(
     page_title="TamamTisch",
