+++ views/about_us.py
@@ -1,8 +1,10 @@
 import streamlit as st
 
-from custom_moduls.calender_widget.helper_functions import months_ahead
+from modules.calender_widget.helper_functions import months_ahead
 
 st.write("This is the about us page.")
 
 st.link_button("Google Sheet",st.session_state.ch.spreadsheet_url)
+
+
