+++ views/tamam_calender.py
@@ -1,6 +1,6 @@
 import streamlit as st
 
-import custom_moduls.calender_widget as cw
+import modules.calender_widget as cw
 
 st.write("This is the calender page.")
